refactor Added video out function.
NB The cv2 writer fails without warning, very picky on the array shape,
encoding, etc.
